Fixed requested changes Added plugs for BitConverter ToUInt16, 32 and 64 methods Used ToUInt methods

+++ source/Cosmos.HAL2/BlockDevice/EBR.cs
@@ -41,8 +41,8 @@ namespace Cosmos.HAL.BlockDevice
             }
             else if (xSystemID != 0)
             {
-                UInt32 xStartSector = (uint)BitConverter.ToInt32(aEBR, (int)aLoc + 8);
-                UInt32 xSectorCount = (uint)BitConverter.ToInt32(aEBR, (int)aLoc + 12);
+                UInt32 xStartSector = BitConverter.ToUInt32(aEBR, (int)aLoc + 8);
+                UInt32 xSectorCount = BitConverter.ToUInt32(aEBR, (int)aLoc + 12);
 
                 var xPartInfo = new PartInfo(xSystemID, xStartSector, xSectorCount);
                 Partitions.Add(xPartInfo);
